Working solution, TODO: Improve RMSE.

+++ src/kalman_filter.cpp
@@ -54,7 +54,7 @@ void KalmanFilter::UpdateEKF(const VectorXd &z) {
   MatrixXd K = P_ * Ht * Si;
   //Estimation
   x_ = x_ + (K * y);
-  long x_size = x_.size();
+  int x_size = x_.size();
   MatrixXd I = MatrixXd::Identity(x_size, x_size);
   P_ = (I - K * H_) * P_;
